testing with v9 and in a new world, fixed 1st bug
